feature: native throttle control
This patch starts the work on native throttle control, removing pv. This
allows for real pausing with signals without uncontrolled buffer/output
semantics, hopefully less forks, and less dependencies.

Another advantage is support for SI throttle parameters in bytes or K.

This is a work in progress. Two areas in particular need work:

- throttle loop

If based on sleep, it needs either GNU sleep or the new Bash sleep with
decimals. This last version needs extra work on some OSes to be enabled,
and both version use different parameters, have different limits and
behave differently. The bash plugin is faster but the GNU version
appears to have more range (tbc).

Mechanisms would have to be in place to support these options properly.

Another options are pure Bash loop, or support for other tools like
usleep.

- playback controls

This is not specific to the native throttle control but arised while
developing the feature. The player/sleeper loops either receive the
controls from playback_control() via normal pipe or process
substitution.

In the first case, playback_control() runs in the same shell and has
(especially) direct access to the tty sending the keystrokes. This
simplifies playback_control(), but the player/sleeper loop is on the
subshell side of the pipe, thus it's messy to get $playback out of
there, control flow is also messy. And we all want Messi, not messy :D

In the second case the situation is reverted, the loops are peachy, and
playback_control() has to deal with race conditions by accessing the tty
saved in KV, what happens when it dies, when it doesn't, it's still on a
pipe or not, etc.

Once these two issues are solved, it's probably fine to phase out pv in
favor of the native throttle, since it does present advantages.