improve: use kv for playback control
Instead of dealing with intricate return codes and whatnot, just store
the playback status in kv, under the namespace pc.

This made straightforward to move the pause loop out of view() into a
new pause_loop() function, and exposed the symmetry between that and the
view_file function - renamed to view_loop(). Both are practically the
same, so will be unified.

This change also makes unnecessary the codes/pc namespace, so it gets
removed as well.