improve: integrate sleep binary autodetection
This is an initial implementation: proper binary detection should be
done via 'type sleep' and reflect that in kv. The idea is to provide
support for various sleep/throttle control methods, including bringing
back pv as an option, and letting the user choose a fixed one, or let
bart to autodetect and use the best implementation available.

Move the builtin detection from top level, to init_misc(), which
eventually set script/bash_loadables.

At update_throttle(), poll kv and use the throttle value semantics
according to the sleep version in use.

At the view file inner loop in subshell, poll kv and eventually enable
the builtin sleep. This is the fastest sleep so far, perhaps minus a
purely-theoretical, pure-bash sleep().