feature: poor man's message queue
In order to simplify and optimize the messaging between the script and
growing number of background processes (user input, playback jobs,
status line and whatnot), add a central facility to send/receive
messages.

The underlying mechanism is essentially the same as to has been used so
far: pipes and signals. This time the pipes are handled by mq() and
rmq().

mq() is the frontend to a background process (rmq) maintaining and
coordinating the message queuing and signal dispatch.

add an initialization secction and two test functions to help with
deving