improve(update_throttle): remove iec converstions
update_throttle is a slow pig in the middle of the whole playback
kerfuffle, spewing pipes left and right. At the least remove the iec
conversions and just convert once, at conf initialization.

There's two pipes that are hard to deal with: awk (that's not really
that hard) and bc (that's really hard). Still, they need go be dealt
with and removed. In any case, profile the bc code and compare to native
bash fp fun, then keep whatever is quicker.

Another strategy is to have a precomputed table with rate scales, then
use that, either for everything or just for big steps. That would
probably speed up things. The goal is to optimize playback loops as
completely as it is possible.